requirements.txt
kafka-python==2.0.6
watchdog==3.0.0
python-dotenv==1.0.0
pydantic>=2.6.0,<3.0.0
docling>=2.26.4
openai>=1.65.0
PyPDF2>=3.0.0
